Move email into separate utils folder
