+++ ai_ta_backend/service/export_service.py
@@ -11,7 +11,7 @@ from injector import inject
 from ai_ta_backend.database.aws import AWSStorage
 from ai_ta_backend.database.sql import SQLDatabase
 from ai_ta_backend.service.sentry_service import SentryService
-from ai_ta_backend.utils.emails import send_email
+from ai_ta_backend.utils.email.send_transactional_email import send_email
 
 
 class ExportService:
@@ -226,7 +226,7 @@ class ExportService:
         return {"response": "Error downloading file!"}
     else:
       return {"response": "No data found between the given dates."}
-    
+
     # Encountered pickling error while running the background task. So, moved the function outside the class.
 
 
@@ -296,7 +296,6 @@ def export_data_in_bg(response, download_type, course_name, s3_path):
     # generate presigned URL
     s3_url = s3.generatePresignedUrl('get_object', os.environ['S3_BUCKET_NAME'], s3_path, 172800)
 
-
     # get admin email IDs
     headers = {"Authorization": f"Bearer {os.environ['VERCEL_READ_ONLY_API_KEY']}", "Content-Type": "application/json"}
 
@@ -338,6 +337,7 @@ def export_data_in_bg(response, download_type, course_name, s3_path):
   except Exception as e:
     print(e)
     return "Error: " + str(e)
+
 
 def export_data_in_bg_emails(response, download_type, course_name, s3_path, emails):
   """
@@ -407,7 +407,7 @@ def export_data_in_bg_emails(response, download_type, course_name, s3_path, emai
 
     admin_emails = emails
     bcc_emails = []
-    
+
     print("admin_emails: ", admin_emails)
     print("bcc_emails: ", bcc_emails)
