+++ ai_ta_backend/utils/email/send_newsletter_email.py
